refactor: move type files back
