+++ packages/superset-ui-chart/src/query/convertMetric.ts
@@ -1,6 +1,6 @@
 import { ChartFormDataMetric } from '../types/ChartFormData';
 import { QueryObjectMetric } from '../types/Query';
-import { AdhocMetric } from '../types/formData/Metric';
+import { AdhocMetric } from '../types/Metric';
 
 export const LABEL_MAX_LENGTH = 43;
