+++ packages/superset-ui-chart/src/query/processMetrics.ts
@@ -1,6 +1,6 @@
 import { ChartFormData } from '../types/ChartFormData';
 import { QueryObjectMetric } from '../types/Query';
-import { MetricKey } from '../types/formData/Metric';
+import { MetricKey } from '../types/Metric';
 import convertMetric from './convertMetric';
 
 export default function processMetrics(formData: ChartFormData) {
