+++ packages/superset-ui-chart/src/types/Annotation.ts
